fix: 5 bugs - add RentPayment model, fix rent_due_day type, fix rent API status check, fix Decimal serialization, add Suspense to profile-setup

+++ backend/app/models/pg_listing.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, JSON, Numeric, String, Text, Uuid, func
+from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, Integer, JSON, Numeric, String, Text, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -25,7 +25,7 @@ class PGListing(Base):
     images: Mapped[list[str] | None] = mapped_column(JSON, nullable=True)
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
     gender_preference: Mapped[GenderPreference | None] = mapped_column(Enum(GenderPreference), nullable=True)
-    rent_due_day: Mapped[int | None] = mapped_column(Numeric(2, 0), nullable=True) # 1-28
+    rent_due_day: Mapped[int | None] = mapped_column(Integer, nullable=True) # 1-28
 
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
